Refactor main.dart and favorites_screen.dart for improved UI and functionality: Simplified main.dart by removing unused imports and classes, and integrated home_screen. Enhanced favorites_screen.dart with modern design elements, improved spacing, and updated card layouts for better user experience. Added new methods for modern detail and WiFi information display.
